requirements.txt
Flask==2.3.2
toml==0.10.2
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==3.0.10